Finished XLTEKStudyFrame and started creating test file.
