--- src/studies/directorytimeframe/directorydayframe.py
@@ -1,37 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-""" directorydayframe.py
-Description:
-"""
-__author__ = "Anthony Fong"
-__copyright__ = "Copyright 2021, Anthony Fong"
-__credits__ = ["Anthony Fong"]
-__license__ = ""
-__version__ = "1.0.0"
-__maintainer__ = "Anthony Fong"
-__email__ = ""
-__status__ = "Prototype"
-
-# Default Libraries #
-
-# Downloaded Libraries #
-
-# Local Libraries #
-from .directorytimeframe import DirectoryTimeFrame
-
-
-# Definitions #
-# Classes #
-# Todo: Maybe skip to XLTEKDayFrame
-class DirectoryDayFrame(DirectoryTimeFrame):
-    """
-
-    Class Attributes:
-
-    Attributes:
-
-    Args:
-
-    """
-
-
